+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Time is now based on precise datetime stamps (specifically, `pd.Timestamp`). In 
 - `ss.Time()` is now called `ss.Timeline()` and its internal calculations are handled differently.
 - `ss.time_ratio()` has been removed; time unit ratio calculations (e.g. months to years) are now handled internally by timepars.
 - `t.abstvec` has been removed; in most cases, `t.tvec` should be used instead (although `t.yearvec`, `t.datevec`, or `t.timevec` may be preferable in some cases).
-- Multiplication by `dt` no longer happens automatically; call `to_prob()` or `p()` to convert from a timepar to a unitless quantity (or `to_events()` or `n()` to convert to a number of events instead).
+- Multiplication by `dt` no longer happens automatically; call `to_prob()` to convert from a timepar to a unitless probability (or `to_events()` to convert to a number of events instead).
 
 For full details, see the migration guide.